add grunt migrate to start script
Updating the container from sharelatex 0.1.4 to sharelatex 0.2.0 fails if grunt migrate script is not called. This is because the datastructures have to be modified for the newer version.

adding migrate command to each start of sharelatex fixes this issue. calling the command multiple times should not be a problem as the script exits with 'nothing to migrate' the second time it is called.