Near complete rework of the Dockerfile. The old one was not working anymore -
I set most things to follow the official container as close as possible.
I tested this container against a data directory from the original and
everything works just flawlessly. In contrast to the official one we
include the full Latex Distribution to make things like gbrief
available.

Due to major changes in Sharelatex we need to use a new node Version and
include several additional packages.

I am using this in production now.